NMS-9288: Topology on Map is not properly displayed
static getHumanReadableIfSpeed moved to InetAddressUtils
